refactor(main-page-layout): move tabular-main-page to main-pages group
