+++ packages/application-components/src/components/main-pages/tabular-main-page/index.ts
